refactor: rename mindmap_orgmode_lists to orgmode_lists
