+++ tests/MyTestCase.test_mindmap_orgmode_lists.approved.txt
@@ -1,4 +1,4 @@
-Executed command: python3 main.py --input ./data/FreePlane/orgmode_lists/input.mm --formatter mindmap_orgmode_lists.py
+Executed command: python3 main.py --input ./data/FreePlane/orgmode_lists/input.mm --formatter orgmode_lists.py
 Result code: 0
 Standard Output (starting on the new line):
 #+title: Export
